+++ sixel.c
@@ -18,7 +18,8 @@ static size_t count;
 void
 sixel_fini(struct terminal *term)
 {
-    free(term->sixel.palette);
+    free(term->sixel.private_palette);
+    free(term->sixel.shared_palette);
 }
 
 static uint32_t
@@ -46,17 +47,36 @@ sixel_init(struct terminal *term)
     term->sixel.image.height = 6;
     term->sixel.image.autosize = true;
 
-    if (term->sixel.palette == NULL) {
-        term->sixel.palette = xcalloc(
-            term->sixel.palette_size, sizeof(term->sixel.palette[0]));
+    /* TODO: default palette */
+
+    if (term->sixel.use_private_palette) {
+        if (term->sixel.private_palette == NULL) {
+            term->sixel.private_palette = xcalloc(
+                term->sixel.palette_size, sizeof(term->sixel.private_palette[0]));
+        } else {
+            /* Private palette - i.e. each sixel starts with a clean palette */
+            memset(
+                term->sixel.private_palette,
+                0,
+                term->sixel.palette_size * sizeof(term->sixel.private_palette[0]));
+        }
+
+        term->sixel.palette = term->sixel.private_palette;
+    } else {
+        if (term->sixel.shared_palette == NULL) {
+            term->sixel.shared_palette = xcalloc(
+                term->sixel.palette_size, sizeof(term->sixel.shared_palette[0]));
+        } else {
+            /* Shared palette - do *not* reset palette for new sixels */
+        }
+
+        term->sixel.palette = term->sixel.shared_palette;
     }
 
     for (size_t i = 0; i < 1 * 6; i++)
         term->sixel.image.data[i] = color_with_alpha(term, term->colors.bg);
 
     count = 0;
-
-    /* TODO: default palette */
 }
 
 void